requirements.txt
Flask
gunicorn
python-dotenv
numpy 
pandas
seaborn 
sklearn
dash
dash_core_components
dash_html_components
keras
tensorflow
schedule
polygon-api-client
websocket-client